+++ console/src/components/Services/Data/TablePermissions/utils.tsx
@@ -214,3 +214,55 @@ export const getPermissionsIcon = (
 
   return 'fullAccess';
 };
+
+export const hasSelectPrimaryKey = (
+  primaryKeys: string[],
+  selectedColumns: string[]
+) =>
+  primaryKeys?.every(primaryKeyColumn =>
+    selectedColumns?.includes(primaryKeyColumn)
+  );
+
+export type Schema = { table_name: string; primary_key: { columns: string[] } };
+export type State = {
+  permissionsState: {
+    table: string;
+  };
+};
+
+export const getPrimaryKeysFromTable = (
+  schemas: Schema[],
+  state: State
+): string[] => {
+  const desiredSchema = schemas?.find(
+    schema => schema.table_name === state.permissionsState.table
+  );
+  return desiredSchema?.primary_key?.columns || [];
+};
+
+export const hasSelectedTablePrimaryKeyFromMetadata = (metadata: any) => {
+  const currentlyEnabledColumns =
+    metadata.permissionsState?.select?.columns || [];
+
+  const currentTablePrimaryKeys = getPrimaryKeysFromTable(
+    metadata.allSchemas,
+    metadata
+  );
+  return hasSelectPrimaryKey(currentTablePrimaryKeys, currentlyEnabledColumns);
+};
+
+type AlteredPermission = 'select_by_pk' | 'select_aggregate';
+type SelectPermission = 'select' | 'select_aggregate' | 'select_by_pk';
+
+export const getNewRootPermissionState = (
+  currentPermissions: SelectPermission[],
+  newPermission: AlteredPermission,
+  isQueryAllowed: boolean
+) => {
+  if (!isQueryAllowed && currentPermissions?.includes(newPermission)) {
+    return currentPermissions.filter(
+      rootPermission => rootPermission !== newPermission
+    );
+  }
+  return currentPermissions;
+};